Saner search function, Fix bonked index page on login

+++ webstore/storefront/models.py
@@ -76,9 +76,9 @@ class Product_Images(models.Model):
 class Listing(models.Model):
     name  = models.CharField(max_length=80)
     inventory = models.IntegerField(default=0)
-    min_price = models.DecimalField(max_digits=7, decimal_places=2)
-    max_price = models.DecimalField(max_digits=7, decimal_places=2)
-    current_price = models.DecimalField(max_digits=7, decimal_places=2)
+    min_price = models.IntegerField()
+    max_price = models.IntegerField()
+    current_price = models.IntegerField()
     rating = models.PositiveIntegerField(null=False, blank=False, default=2.5)
 
     strategy = models.FloatField(default=0.0)
